# [v10.2.x] Search: Modify query for better performance