Backport f999fe3d12cecb4e5be462cea823ffd17d9a2582 from #77576

---

<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

7. If your PR content should be added to the What's New document for the next major or minor release, add the **add to what's new** label to your PR. Note that you should add this label to the main PR that introduces the feature; do not add this label to smaller PRs for the feature.

-->

**What is this feature?**

A customer has reported that they managed to improve search performance in their instance by introducing an index on `dashboard.uid`.
After investigation we realised that the `UQE_dashboard_org_id_uid` was not used.
This fix modifies the query to add `org_id` in the condition so that the above index is used.
It re-introduces also benchmarking [accessControlDashboardPermissionFilter](https://github.com/grafana/grafana/blob/f88a0f36ecf16a88124759b22367f9b50fbecad1/pkg/services/sqlstore/permissions/dashboard.go#L26) permissions filter (in addition to the experimental [accessControlDashboardPermissionFilterNoFolderSubquery](https://github.com/grafana/grafana/blob/f88a0f36ecf16a88124759b22367f9b50fbecad1/pkg/services/sqlstore/permissions/dashboard_filter_no_subquery.go#L15) one)

**Why do we need this feature?**

<details><summary>Execution plan for the query before the fix</summary>
<p>

```
mysql> EXPLAIN SELECT
    ->             dashboard.id,
    ->             dashboard.uid,
    ->             dashboard.title,
    ->             dashboard.slug,
    ->             dashboard_tag.term,
    ->             dashboard.is_folder,
    ->             dashboard.folder_id,
    ->             folder.uid AS folder_uid,
    ->             folder.slug AS folder_slug,
    ->             folder.title AS folder_title  FROM ( SELECT dashboard.id FROM dashboard WHERE ((dashboard.uid IN (SELECT substr(scope, 16) FROM permission WHERE scope LIKE 'dashboards:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
    ->             SELECT ur.role_id
    ->             FROM user_role AS ur
    ->             WHERE ur.user_id = 4816
    ->             AND (ur.org_id = 1 OR ur.org_id = 0)
    ->         UNION
    ->             SELECT br.role_id FROM builtin_role AS br
    ->             WHERE br.role IN ("Editor")
    ->             AND (br.org_id = 1 OR br.org_id = 0)
    ->         ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read") AND NOT dashboard.is_folder) OR (dashboard.folder_id IN (SELECT id FROM dashboard as d WHERE d.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%'  AND role_id IN(SELECT id FROM role INNER JOIN (
    ->             SELECT ur.role_id
    ->             FROM user_role AS ur
    ->             WHERE ur.user_id = 4816
    ->             AND (ur.org_id = 1 OR ur.org_id = 0)
    ->         UNION
    ->             SELECT br.role_id FROM builtin_role AS br
    ->             WHERE br.role IN ("Editor")
    ->             AND (br.org_id = 1 OR br.org_id = 0)
    ->         ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read")) AND NOT dashboard.is_folder) OR (dashboard.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
    ->             SELECT ur.role_id
    ->             FROM user_role AS ur
    ->             WHERE ur.user_id = 4816
    ->             AND (ur.org_id = 1 OR ur.org_id = 0)
    ->         UNION
    ->             SELECT br.role_id FROM builtin_role AS br
    ->             WHERE br.role IN ("Editor")
    ->             AND (br.org_id = 1 OR br.org_id = 0)
    ->         ) as all_role ON role.id = all_role.role_id)  AND action = "folders:read") AND dashboard.is_folder)) AND dashboard.org_id=1 AND dashboard.title LIKE "%genius%" ORDER BY dashboard.title ASC LIMIT 1000 OFFSET 0) AS ids
    ->         INNER JOIN dashboard ON ids.id = dashboard.id
    -> LEFT OUTER JOIN dashboard AS folder ON folder.id = dashboard.folder_id
    ->         LEFT OUTER JOIN dashboard_tag ON dashboard.id = dashboard_tag.dashboard_id
    -> ORDER BY dashboard.title ASC;
+----+--------------+---------------+------------+--------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------+---------+-----------------------------+------+----------+---------------------------------------------------------+
| id | select_type  | table         | partitions | type   | possible_keys                                                                                                                                                                   | key                                  | key_len | ref                         | rows | filtered | Extra                                                   |
+----+--------------+---------------+------------+--------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------+---------+-----------------------------+------+----------+---------------------------------------------------------+
|  1 | PRIMARY      | <derived2>    | NULL       | ALL    | NULL                                                                                                                                                                            | NULL                                 | NULL    | NULL                        |   15 |   100.00 | Using temporary; Using filesort                         |
|  1 | PRIMARY      | dashboard     | NULL       | eq_ref | PRIMARY                                                                                                                                                                         | PRIMARY                              | 8       | ids.id                      |    1 |   100.00 | NULL                                                    |
|  1 | PRIMARY      | folder        | NULL       | eq_ref | PRIMARY                                                                                                                                                                         | PRIMARY                              | 8       | grafana.dashboard.folder_id |    1 |   100.00 | NULL                                                    |
|  1 | PRIMARY      | dashboard_tag | NULL       | ref    | IDX_dashboard_tag_dashboard_id                                                                                                                                                  | IDX_dashboard_tag_dashboard_id       | 8       | ids.id                      |    2 |   100.00 | NULL                                                    |
|  2 | DERIVED      | dashboard     | NULL       | ref    | UQE_dashboard_org_id_folder_id_title,UQE_dashboard_org_id_uid,UQE_dashboard_org_id_folder_uid_title,IDX_dashboard_org_id,IDX_dashboard_org_id_plugin_id,IDX_dashboard_is_folder | UQE_dashboard_org_id_folder_id_title | 8       | const                       |  141 |    10.83 | Using index condition; Using where; Using filesort      |
| 14 | SUBQUERY     | role          | NULL       | index  | PRIMARY                                                                                                                                                                         | IDX_role_org_id                      | 8       | NULL                        |    3 |   100.00 | Using index                                             |
| 14 | SUBQUERY     | <derived16>   | NULL       | eq_ref | <auto_distinct_key>                                                                                                                                                             | <auto_distinct_key>                  | 8       | grafana.role.id             |    1 |   100.00 | Using index                                             |
| 14 | SUBQUERY     | permission    | NULL       | ref    | UQE_permission_role_id_action_scope,IDX_permission_role_id                                                                                                                      | UQE_permission_role_id_action_scope  | 770     | grafana.role.id,const       |   12 |    11.11 | Using where; Using index                                |
| 16 | DERIVED      | ur            | NULL       | ref    | UQE_user_role_org_id_user_id_role_id,IDX_user_role_org_id,IDX_user_role_user_id                                                                                                 | IDX_user_role_user_id                | 8       | const                       |    1 |   100.00 | Using where                                             |
| 17 | UNION        | br            | NULL       | ref    | UQE_builtin_role_org_id_role_id_role,IDX_builtin_role_role,IDX_builtin_role_org_id                                                                                              | IDX_builtin_role_role                | 762     | const                       |    1 |   100.00 | Using where                                             |
| 18 | UNION RESULT | <union16,17>  | NULL       | ALL    | NULL                                                                                                                                                                            | NULL                                 | NULL    | NULL                        | NULL |     NULL | Using temporary                                         |
|  8 | SUBQUERY     | <subquery9>   | NULL       | ALL    | NULL                                                                                                                                                                            | NULL                                 | NULL    | NULL                        | NULL |   100.00 | NULL                                                    |
|  8 | SUBQUERY     | d             | NULL       | index  | PRIMARY                                                                                                                                                                         | UQE_dashboard_org_id_uid             | 171     | NULL                        |  113 |    10.00 | Using where; Using index; Using join buffer (hash join) |
|  9 | MATERIALIZED | role          | NULL       | index  | PRIMARY                                                                                                                                                                         | IDX_role_org_id                      | 8       | NULL                        |    3 |   100.00 | Using index                                             |
|  9 | MATERIALIZED | <derived11>   | NULL       | eq_ref | <auto_distinct_key>                                                                                                                                                             | <auto_distinct_key>                  | 8       | grafana.role.id             |    1 |   100.00 | Using index                                             |
|  9 | MATERIALIZED | permission    | NULL       | ref    | UQE_permission_role_id_action_scope,IDX_permission_role_id                                                                                                                      | UQE_permission_role_id_action_scope  | 770     | grafana.role.id,const       |   12 |    11.11 | Using where; Using index                                |
| 11 | DERIVED      | ur            | NULL       | ref    | UQE_user_role_org_id_user_id_role_id,IDX_user_role_org_id,IDX_user_role_user_id                                                                                                 | IDX_user_role_user_id                | 8       | const                       |    1 |   100.00 | Using where                                             |
| 12 | UNION        | br            | NULL       | ref    | UQE_builtin_role_org_id_role_id_role,IDX_builtin_role_role,IDX_builtin_role_org_id                                                                                              | IDX_builtin_role_role                | 762     | const                       |    1 |   100.00 | Using where                                             |
| 13 | UNION RESULT | <union11,12>  | NULL       | ALL    | NULL                                                                                                                                                                            | NULL                                 | NULL    | NULL                        | NULL |     NULL | Using temporary                                         |
|  3 | SUBQUERY     | role          | NULL       | index  | PRIMARY                                                                                                                                                                         | IDX_role_org_id                      | 8       | NULL                        |    3 |   100.00 | Using index                                             |
|  3 | SUBQUERY     | <derived5>    | NULL       | eq_ref | <auto_distinct_key>                                                                                                                                                             | <auto_distinct_key>                  | 8       | grafana.role.id             |    1 |   100.00 | Using index                                             |
|  3 | SUBQUERY     | permission    | NULL       | ref    | UQE_permission_role_id_action_scope,IDX_permission_role_id                                                                                                                      | UQE_permission_role_id_action_scope  | 770     | grafana.role.id,const       |   12 |    11.11 | Using where; Using index                                |
|  5 | DERIVED      | ur            | NULL       | ref    | UQE_user_role_org_id_user_id_role_id,IDX_user_role_org_id,IDX_user_role_user_id                                                                                                 | IDX_user_role_user_id                | 8       | const                       |    1 |   100.00 | Using where                                             |
|  6 | UNION        | br            | NULL       | ref    | UQE_builtin_role_org_id_role_id_role,IDX_builtin_role_role,IDX_builtin_role_org_id                                                                                              | IDX_builtin_role_role                | 762     | const                       |    1 |   100.00 | Using where                                             |
|  7 | UNION RESULT | <union5,6>    | NULL       | ALL    | NULL                                                                                                                                                                            | NULL                                 | NULL    | NULL                        | NULL |     NULL | Using temporary                                         |
+----+--------------+---------------+------------+--------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------+---------+-----------------------------+------+----------+---------------------------------------------------------+
25 rows in set, 1 warning (0.02 sec)
```

</p>
</details> 

<details><summary>Execution plan the query after the fix</summary>
<p>

```
mysql> EXPLAIN SELECT
    ->                         dashboard.id,
    ->                         dashboard.uid,
    ->                         dashboard.title,
    ->                         dashboard.slug,
    ->                         dashboard_tag.term,
    ->                         dashboard.is_folder,
    ->                         dashboard.folder_id,
    ->                         folder.uid AS folder_uid,
    -> 
    ->                         folder.slug AS folder_slug,
    ->                         folder.title AS folder_title  FROM ( SELECT dashboard.id FROM dashboard WHERE ((dashboard.uid IN (SELECT substr(scope, 16) FROM permission WHERE scope LIKE 'dashboards:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
    ->                         SELECT ur.role_id
    ->                         FROM user_role AS ur
    ->                         WHERE ur.user_id = 4816
    ->                         AND (ur.org_id = 1 OR ur.org_id = 0)
    ->                 UNION
    ->                         SELECT br.role_id FROM builtin_role AS br
    ->                         WHERE br.role IN ("Editor")
    ->                         AND (br.org_id = 1 OR br.org_id = 0)
    ->                 ) as all_role ON role.id = all_role.role_id)  AND action =  "dashboards:read") AND NOT dashboard.is_folder) OR (dashboard.folder_id IN (SELECT d.id FROM dashboard as d WHERE d.org_id = 1 AND d.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%'  AND role_id IN(SELECT id FROM role INNER JOIN (
    ->                         SELECT ur.role_id
    ->                         FROM user_role AS ur
    ->                         WHERE ur.user_id = 4816
    ->                         AND (ur.org_id = 1 OR ur.org_id = 0)
    ->                 UNION
    ->                         SELECT br.role_id FROM builtin_role AS br
    ->                         WHERE br.role IN ("Editor")
    ->                         AND (br.org_id = 1 OR br.org_id = 0)
    ->                 ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read")) AND NOT dashboard.is_folder) OR (dashboard.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
    ->                         SELECT ur.role_id
    ->                         FROM user_role AS ur
    ->                         WHERE ur.user_id = 4816
    ->                         AND (ur.org_id = 1 OR ur.org_id = 0)
    ->                 UNION
    ->                         SELECT br.role_id FROM builtin_role AS br
    ->                         WHERE br.role IN ("Editor")
    ->                         AND (br.org_id = 1 OR br.org_id = 0)
    ->                 ) as all_role ON role.id = all_role.role_id)  AND action = "folders:read") AND dashboard.is_folder)) AND dashboard.org_id=1 AND dashboard.title LIKE "%genius%" ORDER BY dashboard.title ASC LIMIT 1000 OFFSET 0) AS ids
    ->                 INNER JOIN dashboard ON ids.id = dashboard.id
    -> 
    ->                 LEFT OUTER JOIN dashboard AS folder ON folder.id = dashboard.folder_id
    ->         LEFT OUTER JOIN dashboard_tag ON dashboard.id = dashboard_tag.dashboard_id
…